chore: remove deprecated agents package
