--- packages/agents/package.json
@@ -1,51 +1,0 @@
-{
-  "name": "@cairo-coder/agents",
-  "version": "1.0.0",
-  "private": true,
-  "scripts": {
-    "build": "tsc -p tsconfig.json",
-    "test": "jest",
-    "test-doc-quality": "ts-node src/testDocQuality.ts"
-  },
-  "dependencies": {
-    "@iarna/toml": "^2.2.5",
-    "@langchain/anthropic": "^0.2.18",
-    "@langchain/community": "^0.3.32",
-    "@langchain/core": "^0.2.36",
-    "@langchain/google-genai": "^0.1.8",
-    "@langchain/mongodb": "^0.0.5",
-    "@langchain/openai": "^0.0.25",
-    "@types/node": "^20",
-    "commander": "^11.1.0",
-    "compute-cosine-similarity": "^1.1.0",
-    "compute-dot": "^1.1.0",
-    "dotenv": "^16.4.7",
-    "mongodb": "^6.13.1",
-    "node-fetch": "^3.3.2",
-    "pg": "^8.14.1",
-    "winston": "^3.17.0"
-  },
-  "exports": {
-    ".": {
-      "types": "./src/index.ts",
-      "default": "./dist/index.js"
-    },
-    "./*": {
-      "types": "./src/*.ts",
-      "default": "./dist/*.js"
-    }
-  },
-  "devDependencies": {
-    "@cairo-coder/typescript-config": "workspace:*",
-    "@types/jest": "^29.5.14",
-    "@types/supertest": "^6.0.2",
-    "jest": "^29.7.0",
-    "jest-mock-extended": "4.0.0-beta1",
-    "nodemon": "^3.1.9",
-    "prettier": "^3.5.2",
-    "supertest": "^7.0.0",
-    "ts-jest": "^29.2.5",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.7.3"
-  }
-}
