--- packages/agents/src/index.ts
@@ -1,7 +1,0 @@
-export * from './types';
-export * from './config/settings';
-export * from './config/agent';
-export * from './config/agents';
-export * from './core/agentFactory';
-export * from './db/postgresVectorStore';
-export * from './utils';
