+++ packages/ingester/package.json
@@ -10,7 +10,6 @@
   "dependencies": {
     "@antora/lunr-extension": "1.0.0-alpha.10",
     "@asciidoctor/tabs": "1.0.0-beta.6",
-    "@cairo-coder/agents": "workspace:*",
     "@iarna/toml": "^2.2.5",
     "@langchain/core": "^0.2.36",
     "@langchain/openai": "^0.0.25",
@@ -20,6 +19,7 @@
     "dotenv": "^16.4.7",
     "downdoc": "1.0.2-stable",
     "lunr": "^2.3.9",
+    "pg": "^8.14.1",
     "winston": "^3.17.0"
   },
   "devDependencies": {
