+++ packages/ingester/src/ingesters/OpenZeppelinDocsIngester.ts
@@ -1,11 +1,11 @@
 import * as fs from 'fs/promises';
 import * as path from 'path';
 import { Document } from '@langchain/core/documents';
-import { BookChunk, DocumentSource } from '@cairo-coder/agents/types/index';
+import { BookChunk, DocumentSource } from '../types';
 import { BookConfig } from '../utils/types';
-import { logger } from '@cairo-coder/agents/utils/index';
+import { logger } from '../utils/logger';
 import { MarkdownIngester } from './MarkdownIngester';
-import { VectorStore } from '@cairo-coder/agents/db/postgresVectorStore';
+import { VectorStore } from '../db/postgresVectorStore';
 import { calculateHash } from '../utils/contentUtils';
 import {
   RecursiveMarkdownSplitter,
