+++ packages/ingester/src/ingesters/StarknetDocsIngester.ts
@@ -1,8 +1,8 @@
 import * as path from 'path';
 import { Document } from '@langchain/core/documents';
-import { BookChunk, DocumentSource } from '@cairo-coder/agents/types/index';
+import { BookChunk, DocumentSource } from '../types';
 import { BookConfig, BookPageDto, ParsedSection } from '../utils/types';
-import { logger } from '@cairo-coder/agents/utils/index';
+import { logger } from '../utils/logger';
 import { exec as execCallback } from 'child_process';
 import { promisify } from 'util';
 import { MarkdownIngester } from './MarkdownIngester';
