+++ packages/ingester/src/ingesters/StarknetFoundryIngester.ts
@@ -2,10 +2,10 @@ import * as fs from 'fs/promises';
 import * as path from 'path';
 import axios from 'axios';
 import AdmZip from 'adm-zip';
-import { BookChunk, DocumentSource } from '@cairo-coder/agents/types/index';
+import { BookChunk, DocumentSource } from '../types';
 import { BookConfig, BookPageDto, ParsedSection } from '../utils/types';
 import { processDocFiles } from '../utils/fileUtils';
-import { logger } from '@cairo-coder/agents/utils/index';
+import { logger } from '../utils/logger';
 import { exec as execCallback } from 'child_process';
 import { promisify } from 'util';
 import { MarkdownIngester } from './MarkdownIngester';
